before a big change
